+++ src/components/Tabs.tsx
@@ -126,7 +126,6 @@ const TabBtn = styled('button')<{
 
 const Panel = styled('div')`
   padding: 1rem 0;
-  overflow: hidden;
 `;
 
 /*───────────────────────────────────────────────────────────*/