+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file. The format 
 - Stack default padding / margins
 ### Changed
 - `Table` now defaults to striped rows and column dividers
+### Fixed
+- Tab panels no longer create stray scrollbars when tables are constrained
 
 ## [v0.7.2]
 ### Changed